cStanzaDataStructures.EcopathCode should be integer Added sponsor to merge groups UI Fixed compatible group detection
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@14436 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwECore/MultiStanza/cStanzaDatastructures.vb
@@ -127,6 +127,7 @@ Public Class cStanzaDatastructures
 
     ''' <summary>Egg production shape is seasonal.</summary>
     Public EggProdIsSeasonal() As Boolean
+
     ''' <summary></summary>
     ''' <remarks>
     ''' <list type="bullet">
@@ -135,7 +136,7 @@ Public Class cStanzaDatastructures
     ''' <item>2: Ecopath no for leading QB stanza.</item>
     ''' </list>
     ''' </remarks>
-    Public SpeciesCode(,) As Single
+    Public SpeciesCode(,) As Integer
 
 #Region " Private data "
